Introduced primary adapter.
