+++ patterns/design-patterns-architectural/src/main/java/com/baeldung/hexagonal/adapter/SecondaryRegularDiscountCalculator.java
@@ -2,11 +2,11 @@ package com.baeldung.hexagonal.adapter;
 
 import com.baeldung.hexagonal.port.DiscountCalculator;
 
-public class SalesDiscountCalculator implements DiscountCalculator {
+public class SecondaryRegularDiscountCalculator implements DiscountCalculator {
 
     @Override
     public float applyDiscount(float price) {
-        return price * .75F;
+        return price * .95F;
     }
 
 }
