+++ patterns/design-patterns-architectural/src/main/java/com/baeldung/hexagonal/adapter/SecondarySalesDiscountCalculator.java
@@ -2,11 +2,11 @@ package com.baeldung.hexagonal.adapter;
 
 import com.baeldung.hexagonal.port.DiscountCalculator;
 
-public class RegularDiscountCalculator implements DiscountCalculator {
+public class SecondarySalesDiscountCalculator implements DiscountCalculator {
 
     @Override
     public float applyDiscount(float price) {
-        return price * .95F;
+        return price * .75F;
     }
 
 }
