+++ patterns/design-patterns-architectural/src/main/java/com/baeldung/hexagonal/application/Application.java
@@ -1,20 +1,18 @@
 package com.baeldung.hexagonal.application;
 
 import com.baeldung.hexagonal.port.PriceCalculator;
-import com.baeldung.hexagonal.port.DiscountCalculator;
 
-public class Application implements PriceCalculator {
-    private final DiscountCalculator provider;
+public class Application {
+    private final PriceCalculator primaryAdapter;
 
-    public Application(DiscountCalculator provider) {
-        this.provider = provider;
+    public Application(PriceCalculator primaryAdapter) {
+        super();
+        this.primaryAdapter = primaryAdapter;
     }
 
-    @Override
-    public float calculateFinalPrice(float shoppingCartPrice) {
-        float priceBeforeShipping = this.provider.applyDiscount(shoppingCartPrice);
-        float finalPrice = priceBeforeShipping < 40F ? priceBeforeShipping + 3.99F : priceBeforeShipping;
-
-        return finalPrice;
+    public PriceCalculator getPrimaryAdapter() {
+        return primaryAdapter;
     }
+
+   
 }
